server/mcp_server_bytehouse/pyproject.toml
[project]
name = "mcp_bytehouse"
version = "0.1.0"
description = "An MCP server for ByteHouse."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
     "mcp[cli]>=1.6.0",
     "python-dotenv>=1.0.1",
     "uvicorn>=0.34.0",
     "clickhouse-connect>=0.8.0",
     "pip-system-certs>=4.0",
]

[project.scripts]
mcp_bytehouse = "mcp_bytehouse.main:main"


[project.optional-dependencies]
dev = [
    "ruff",
    "pytest"
]


[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_bytehouse"]
